Fix invalid use of Selector and SelectionKey classes. No longer freezes after channel change.

+++ src/fi_81/cwp_morse_mangle/cwp/CWFrequencyChange.java
@@ -2,10 +2,18 @@ package fi_81.cwp_morse_mangle.cwp;
 
 public class CWFrequencyChange extends CWStateChange {
 	public CWFrequencyChange(long newFreq) {
+		/*
+		 * Register with correct type and add timestamp of 0 (making sending
+		 * logic to pass this forward immediately.
+		 */
 		super(CWStateChange.TYPE_FREQUENCY_CHANGE, (int) (-newFreq), 0);
 	}
 
 	public long getFrequency() {
-		return -value;
+		/*
+		 * Must cast to 'long' before negation, integer negation of
+		 * Integer.MIN_VALUE results overflow back to Integer.MIN_VALUE.
+		 */
+		return -(long) value;
 	}
 }
